Moving the renderer logic
